+++ RayTracer/src/Vector3.h
@@ -6,49 +6,49 @@ namespace RayTracer
 	{
 		union
 		{
-			double X = 0.0;
-			double R;
+			float X = 0.0;
+			float R;
 		};
 		union
 		{
-			double Y = 0.0;
-			double G;
+			float Y = 0.0;
+			float G;
 		};
 		union
 		{
-			double Z = 0.0;
-			double B;
+			float Z = 0.0;
+			float B;
 		};
 
 		Vector3() = default;
-		Vector3(double x, double y, double z);
+		Vector3(float x, float y, float z);
 
 		bool IsNearlyNull() const;
 
-		double GetLength() const;
-		double GetSquaredLength() const;
+		float GetLength() const;
+		float GetSquaredLength() const;
 
 		Vector3 GetUnitVector() const;
 
 		Vector3 Reflect(const Vector3& normal) const;
-		Vector3 Refract(const Vector3& unitNormal, double refractiveIndicesRatio) const;
+		Vector3 Refract(const Vector3& unitNormal, float refractiveIndicesRatio) const;
 
-		Vector3 Rotate(double angle, const Vector3& axis) const;
+		Vector3 Rotate(float angle, const Vector3& axis) const;
 
 		Vector3& operator+=(const Vector3& v);
 		Vector3& operator-=(const Vector3& v);
-		Vector3& operator*=(double t);
-		Vector3& operator/=(double t);
+		Vector3& operator*=(float t);
+		Vector3& operator/=(float t);
 
 		friend Vector3 operator+(Vector3 v1, const Vector3& v2);
 		friend Vector3 operator-(Vector3 v1, const Vector3& v2);
-		friend Vector3 operator*(Vector3 v, double t);
-		friend Vector3 operator*(double t, Vector3 v);
-		friend Vector3 operator/(Vector3 v, double t);
+		friend Vector3 operator*(Vector3 v, float t);
+		friend Vector3 operator*(float t, Vector3 v);
+		friend Vector3 operator/(Vector3 v, float t);
 
 		friend Vector3 operator-(const Vector3& v);
 
-		friend double Dot(const Vector3& v1, const Vector3& v2);
+		friend float Dot(const Vector3& v1, const Vector3& v2);
 		friend Vector3 Cross(const Vector3& v1, const Vector3& v2);
 	};
